Remove R.java from the repository
Since R.java is a generated file it does not need to recorded in git.
One of the first things that happened when I imported the project in
to eclipse was that it got regenerated and the result was slightly
different than the version that was in the git repository.  This was
probably because my version of the Android SDK is slightly different
than the one that was used when the file was committed.

Rather than thrash on slight variants of this file, it is best not to
record the file at all.  The SDK will generate the file when it needs
it.

Change-Id: I7c4d6ea16d4ceeade57148c6116e2ca8068bc3f1